move method into a more appropriate file
