--- src/utils/Room.ts
@@ -1,38 +1,0 @@
-/*
-Copyright 2021 The Matrix.org Foundation C.I.C.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-import { Room } from "matrix-js-sdk/src/models/room";
-
-import SpaceStore from "../stores/spaces/SpaceStore";
-import DMRoomMap from "./DMRoomMap";
-import { _t } from "../languageHandler";
-
-export function roomContextDetailsText(room: Room): string {
-    if (room.isSpaceRoom()) return undefined;
-
-    const dmPartner = DMRoomMap.shared().getUserIdForRoomId(room.roomId);
-    if (dmPartner) {
-        return room.getMember(dmPartner)?.name;
-    }
-
-    const [parent, ...otherParents] = SpaceStore.instance.getKnownParents(room.roomId);
-    if (parent) {
-        return _t("%(spaceName)s and %(count)s others", {
-            spaceName: room.client.getRoom(parent).name,
-            count: otherParents.length,
-        });
-    }
-}
